rearrange client functions and new DB API
