+++ Makefile.am
@@ -14,7 +14,8 @@ libdqlite_la_CFLAGS = $(AM_CFLAGS) -fvisibility=hidden
 libdqlite_la_LDFLAGS = $(AM_LDFLAGS) -version-info 0:1:0
 libdqlite_la_SOURCES = \
   src/bind.c \
-  src/client/protocol.c \
+  src/client.c \
+  src/client_protocol.c \
   src/command.c \
   src/conn.c \
   src/db.c \
